twitter_clone/server/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
factory_boy==3.3.3
Faker==37.4.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-xdist==3.8.0
pytest-asyncio==1.0.0
pytest-flask==1.3.0
pyzmq==27.0.0
Werkzeug==3.1.3
black==25.1.0
flake8==7.3.0
flake8-bugbear==24.12.12
flake8-pie==0.16.0
isort==6.0.1
mypy==1.16.1
flask_sqlalchemy==3.1.1
types-Flask==1.1.6
sqlalchemy==2.0.41
gunicorn==23.0.0
psycopg2-binary==2.9.10
flasgger==0.9.7.1
flask-postgresql==1.1.1